requirements.txt
requests==2.31.0
openai==1.3.5
gspread==6.0.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1